privatevault-enterprise-pilot/interop/agent-frameworks/requirements.txt
pytest

# Agent frameworks
langchain
langchain-community
langgraph
pyautogen
crewai

# RAG frameworks
llama-index
haystack-ai

# Enterprise / orchestration
semantic-kernel

# Google Agent Development Kit
google-adk
